feat: middleware 내부 로직 함수로 분리
